+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Not Released
 
+- NEW : Apply default price - *09/10/2023* - 2.22.0
 - NEW : TakePos module compatibility - *23/08/2023* - 2.21.0
 
 ## 2.20